Split nanobind wrapper into header and implementation
Follow the DOLFINx convention of keeping the shared helpers and
templates in a header and the module registration in the
implementation file.

python/basix_wrappers/basix_wrappers.h holds the as_nbarray helpers,
dispatch_dtype, cell_type_to_str and the declare_float<T> template.
python/wrapper.cpp is reduced to NB_MODULE(_basixcpp, m).

The anonymous namespace becomes the named basix_wrappers namespace, and
`using namespace basix;` is dropped from the header so that it does not
leak into includers.

The header is installed into the Python package and can be located with
the new basix.get_include(), so that downstream nanobind wrappers can
reuse it.
